Split up utils module functions
