+++ backend/tests/integration/test_api_castling.py
@@ -4,7 +4,7 @@ import pytest
 
 from mocks.starting_game import starting_game
 import src.api as api
-from src.utils import clear_game
+from src.utils.game_state import clear_game
 from tests.test_utils import (
     select_and_move_white_piece, select_and_move_black_piece
 )
