+++ backend/tests/integration/test_api_pawn_mechanics.py
@@ -3,7 +3,7 @@ from fastapi import HTTPException, Response
 import pytest
 
 import src.api as api
-from src.utils import clear_game
+from src.utils.game_state import clear_game
 from tests.test_utils import (
     select_and_move_white_piece, select_and_move_black_piece
 )
